+++ src/clerks_friend/utils.py
@@ -1,17 +1,22 @@
 __all__ = ["is_dataclass_instance", "dataclass_list_to_table"]
 
-from dataclasses import fields, is_dataclass
+import dataclasses
 from typing import Any
 
 from prettytable import MARKDOWN, PrettyTable
 
 
 def is_dataclass_instance(dataclass_instance: Any) -> bool:
-    return is_dataclass(dataclass_instance) and not isinstance(dataclass_instance, type)
+    return dataclasses.is_dataclass(dataclass_instance) and not isinstance(
+        dataclass_instance, type
+    )
 
 
 def dataclass_list_to_table(
-    dataclass_list: list[Any], alignment: str | list[str] = "l"
+    dataclass_list: list[Any],
+    *,
+    alignment: str | list[str] = "l",
+    fields: list[str] | None = None
 ) -> str:
     # The table header will come from the first item in the list. We'll also only do
     # a quick check to make sure this item is a dataclass instance.
@@ -20,10 +25,14 @@ def dataclass_list_to_table(
     if not is_dataclass_instance(dataclass_list[0]):
         raise TypeError("Unsupported data type")
 
-    table_fields = fields(dataclass_list[0])
+    table_fields = dataclasses.fields(dataclass_list[0])
     table = PrettyTable()
     table.set_style(MARKDOWN)
-    field_names = [field.name.replace("_", " ").title() for field in table_fields]
+
+    if fields is None:
+        fields = [field.name for field in table_fields]
+    field_names = [field.replace("_", " ").title() for field in fields]
+
     table.field_names = field_names
     if isinstance(alignment, str):
         table.align = alignment
@@ -31,7 +40,7 @@ def dataclass_list_to_table(
         table.align.update(dict(zip(field_names, alignment)))
 
     for dc in dataclass_list:
-        values = [getattr(dc, x.name) for x in table_fields]
+        values = [getattr(dc, x) for x in fields]
         table.add_row(values)
 
     return table.get_string()