+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Members Moved Out report
+- Ability to specify the fields to keep in the output table
+
 ### Changed
 ### Deprecated
 ### Removed